feat(app): checks if user are admin to allow delete itens and filters extras and block by unit Id

+++ src/services/AllocatableService.ts
@@ -148,8 +148,15 @@ export class AllocatableService {
 
   async delete(
     entityToDelete: number,
-    user?: UserDTO,
+    user: UserDTO,
   ): Promise<number | null | undefined> {
+    if (user.role !== 'admin')
+      throw HttpExceptionDTO.warn(
+        `User not have permission`,
+        'Usuário não tem permissão para deletar recursos',
+        HttpStatus.FORBIDDEN,
+      );
+
     const deletedEntities =
       await this.allocatableRepository.delete(entityToDelete);
     return deletedEntities.affected;
